implement parameterized tests
This allows to execute the same test with different parameters and
treats each execution as an own test. The change consists of two parts,
the TestCaller can now handle any callable which also makes it easy to
generate programatically more complex test cases as well as the new
CPPUNIT_TEST_PARAMETERIZED macro. That macro takes the test name as well
as an iteratable, e.g. std::initializer_list.

An example for this usage is:

class SimpleTest : public CppUnit::TestFixture
{
public:
    CPPUNIT_TEST_SUITE(SimpleTest);
    CPPUNIT_TEST_PARAMETERIZED(test, {1, 2, 3, 4});
    CPPUNIT_TEST_SUITE_END();

    void test(int i)
    {
        CPPUNIT_ASSERT(i < 5);
    }
};

which will execute test 4 times with the values 1 to 4.